Init IAU sdk in app delegate
